Initial, but ready to use version of the log4bsh project.
log4bsh is a configureable logger facility for bash scripts.

Besides different log messages, like info, debug, trace, warning, error,
it offers to override default settings in different ways,
via environment variables, a user config or a systemwide config.

The name of the script that calls the logging functions is written to
the log file, so one known which messages are caused by what  script.

Further, it offers hooks for code to be executed on exit, caused by to
an error message. And to modify the name of scripts printed tothe log,
like for example to remove its file extension.

Change-Id: I5a5702dbf6b5aa399fe5f93a751069aecddaaff8
Signed-off-by: Nico Struckmann <struckmann@hlrs.de>